Space identification fixed
Finally fixed it.

We use checks to see if there is enough space for the object to go on a particular spot. We also add corrected coloring logic, meaning that it'll only color the boxes it's filling, and the others it isn't will return back to its default state. I also resized the images because in this case, they were too large.

Next up, grid and item rotations!