Fixed vertical rotation issue
It works as intended:

- Verticality
- Doesn't anchor to top left of screen
- Accurate grid spacing
- Taken and Free grid states
- Little snap animation
- Item doesn't exceed bounds of grid
- Pickup and Drop added (last commit)

The next thing to do is to improve the UI and make it work in the world scene for testing. A future goal is to get a similar system working on the left side of the screen where the source inventory will be. The right side is the player's inventory.

PS: Also have to add all the other items to the item data, and that takes forever!!!