Adding interaction functionality
When the player walks up to the item, it highlights the item in a subtle green. If it were too far form the item, it will disable the highlight.

This logic isn't implemented with the actual inventory yet, so whenever an actual item is picked up, it doesn't go inside the inventory. This will be stored later when we fix the inventory's drag and drop system.