Drag and drop + Stacking
The items stack and display the correct values now. Before when we would drop an item into a lot, the item's quantity would still be the same when only dropping once. But now when we right click onto an empty slot, it only inserts one item into the slot unless clicked multiple times.

We also introduced spot swapping and a drag and drop system. In the future, we may consider adding colors to indicate if an action performed in the inventory can or cannot be done, like red-light green-light you know?