External inventory fixed
By making the clear_external_inventory function and adding some connections using signals, we can ensure that the external loot source doesn't open along with the regular inventory when pressing the open_inv.

It's done, and now we have to be able to pick up and drop the items out of the inventory!