Testing drop concept using print statements
By using "match" and making a custom signal, for every time the item outside of the inventory is left-clicked, we can use print statements to determine if the drop will function as intended.

It doesn't work yet, but we know it'll work once we add the rest of the code since the print statement went through as "drop data" meaning that when we left-clicked with the item outside the inventory, it dropped.