Consumable class complete
Now we have a check that determines whether or not the item that the user right-clicks on is actually a consumable or not.

If the item is a consumable, then the item disappears and adds health to the player. But if it isn't a consumable, the item doesn't do anything.