Functioning healthbar
It works as expected:

When the player eats a consumable, the player gains health and the healthbar updates itself accordingly.

We made sure the healthbar was below the max to test this function.
It's good, but still deciding if the white bar should be kept or not in the health bar, more to add on later.

💯