The Armor Bar
Pretty much the same logic as the health bar, except the fact that it's armor instead of health.

The cool thing is, because we used a ItemDataConsumable class, we don't really have to change anything about the armor or consumables. All we do is give the armor a value of how much armor it'll increase the bar by, and boom.

💰