Still fixing the weight functionality
So much work needs to be fixed. Still unsure why the weight from the start isn't the weight from the items already in the inventory.

But some stuff was fixed, like when the weight would over exceed the limit, and when you could drag items into slots even when there wasn't enough weight, or no weight at all.

Still though, the initial weight is the issue.