Adding Equipment
A temporary image is there until the actual helmet is made. The equipment class was also made as a resource so that the item contents can be sourced globally (meaning I can access it from anywhere).

It has no effect yet, but the first steps were to get it equipped in a slot, and make sure the image was popping up. We also made it so that its slot doesn't accept items that aren't equipment.