Making the weapon scene
We have a weapon scene, FINALLY.

It doesn't work yet, but we get the weapon to load properly into the test_room and attached to the player. But the weapon doesn't shoot, and it doesn't have any slot data reference in the inventory (meaning the item is technically not there in the inventory).