Making a loading screen, added some animations
Some looping elements were added as well as the loading screen itself. We were able to make it so that the loading text was animated as it was loading.

The loading screen can also be accessed from anywhere meaning that a new loading screen wouldn't have to be made every single time a new scene would load. It would just instantiate itself in the needed time.